requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
selenium==4.26.0

